Show users location options found with their form inputs
The LocationIQ API returns an array of locations when it receives a
request for geocoordinates.  Prior to this commit, the app
automatically selected the first of these locations and ran the rest
of the wildlife search based on that location.  Now the app lets the
user see all the locations returned by LocationIQ, so they can choose
which location best meets their needs.