Make the 'Street Name' field required
The LocationIQ API only returns coordinates for its locations, no
the size or shape.  A large area, like a country, can only be
represented by a box area, and will not be represented with its
actual shape.  Without this change, the user could think they're
searching for wildlife across an entire region, when actually the
search area is an inaccurate representation of the region.  Making
the "Street Name" field required prevents the user from being misled
by the search results of a large area.