Add limits and defaults to start and end dates
The following changes were made:
  - Prevent users from entering future dates
  - Set the default start date to a year before the current date
  - Set the default end date to the current date
  - Change labels for the start and end date to be more clear to user
  - Prevent users from entering a start date that is later then the end
    date

No minimum was set for the start date field. iNaturalist allows users
to add observations that occurred at any time in the past. Instead, my
app makes it clear that it is only searching for records of
observations of wildlife, and not all possible wildlife that exists.